+++ L3G4200D.c
@@ -14,9 +14,9 @@ INT16 L3G4200D_YAngularRate_Raw = 0;
 INT16 L3G4200D_ZAngularRate_Raw = 0;
 INT16 L3G4200D_Temperature_Raw = 0;
 
-INT16 L3G4200D_XAngularRate_Raw_Avg = 0;
-INT16 L3G4200D_YAngularRate_Raw_Avg = 0;
-INT16 L3G4200D_ZAngularRate_Raw_Avg = 0;
+float L3G4200D_XAngularRate_Raw_Avg = 0;
+float L3G4200D_YAngularRate_Raw_Avg = 0;
+float L3G4200D_ZAngularRate_Raw_Avg = 0;
 
 //******************************************************************************
 //Public Function Definitions
